ADC DAC oucart-0020 + lecture et chargement calib json ok
- changment du switch i2c lors du moyennage des valeurs pour pas couper le switch à chaque moyenne
- Ini du cna à 0 désactivée
-Ajout du rack dans simcaptemp
-readadc() msb lsb
-setdac() centrage sur 2.5v
-IDN dans json passe en string
-equation calibration readcurrent()
-json valeur par défaut si exeption
-modifications simcaptemp pour rack

+++ Integrationcppstatemachine/Src/OUCART0020.h
@@ -20,7 +20,7 @@ public:
 	void setI2cAdress(uint8_t adr);
 	uint16_t readADC(uint8_t channel); //lit la voie adc
 
-	uint16_t setDAC(uint8_t channel, uint16_t value); //change la valeur du dac
+	uint16_t setDAC(uint8_t channel, int value); //change la valeur du dac
 
 	uint8_t setconfigADC( uint8_t pins);
 	uint8_t setconfigDAC( uint8_t pins);
